Add more rules to the Ruff linter
This is the result of me working on songbird and taking a look at Ruff's
default config. It turns out I think that there's a lot that can be
enabled!

One exception though is pycodestyle's E501 (line-length): sometimes you
do need big strings and it's a hassle to go around the linter to build
them through concatenation. Ruff format already does a good job at
breaking up function calls and other things to respect the 88
line-length rule.

I expect this to change in the future!